requirements.txt
googlesearch-python 
requests